deconv fixes

+++ src/yass/preprocess/util.py
@@ -38,7 +38,7 @@ def _butterworth(ts, low_frequency, high_factor, order, sampling_frequency):
     low = float(low_frequency) / sampling_frequency * 2
     high = float(high_factor) * 2
     b, a = butter(order, low, btype='high', analog=False)
-
+    
     if ts.ndim == 1:
         return filtfilt(b, a, ts)
     else:
@@ -105,6 +105,10 @@ def _standardize(rec, sd=None, centers=None):
     return rec
     #return np.divide(rec, sd)
 
+def whiten(ts):
+    
+    wrot = np.load('/ssd/nishchal/temp_/new/neuropixels-data-sep-2020/scripts/recordings/wrot.npy')
+    return np.matmul(wrot, ts.T).T
 
 def filter_standardize_batch(batch_id, reader, fname_mean_sd,
                              apply_filter, out_dtype, output_directory,
@@ -136,10 +140,11 @@ def filter_standardize_batch(batch_id, reader, fname_mean_sd,
     
     if os.path.exists(os.path.join(
         output_directory,
-        "standardized_{}.npy".format(
+        "filtered_{}.npy".format(
             str(batch_id).zfill(6)))):
         return
     
+        
     # filter
     if apply_filter:
         # read a batch
@@ -149,6 +154,7 @@ def filter_standardize_batch(batch_id, reader, fname_mean_sd,
         ts = ts[reader.buffer:-reader.buffer]
     else:
         ts = reader.read_data_batch(batch_id, add_buffer=False)
+    
 
     # standardize
     temp = np.load(fname_mean_sd)
@@ -156,19 +162,15 @@ def filter_standardize_batch(batch_id, reader, fname_mean_sd,
     centers = temp['centers']
     ts = _standardize(ts, sd, centers)
     
+    ts = whiten(ts)
+    
     # save
     fname = os.path.join(
         output_directory,
-        "standardized_{}.npy".format(
+        "filtered_{}.npy".format(
             str(batch_id).zfill(6)))
     np.save(fname, ts.astype(out_dtype))
 
-    #fname = os.path.join(
-    #    output_directory,
-    #    "standardized_{}.bin".format(
-    #        str(batch_id).zfill(6)))
-    #f = open(fname, 'wb')
-    #f.write(ts.astype(out_dtype))
 
 
 
@@ -258,7 +260,7 @@ def make_histograms(batch_id, reader, output_directory, output_directory_spikes,
     #### Make histograms ####
     for spike_time in np.where(ptp_sliding.max(0)>=voltage_threshold)[0]:
         electrode_ptp_int = np.log1p(np.matmul(ptp, M))
-        hist_plot = np.histogram2d(electrode_ptp_int, np.arange(0, num_y_pos), bins=(20, num_y_pos), range = [[np.log1p(voltage_threshold), np.log1p(10*voltage_threshold)], [0, num_y_pos]])[0]
+        hist_plot = np.histogram2d(electrode_ptp_int, np.arange(0, num_y_pos), bins=(20, num_y_pos), range = [[np.log1p(voltage_threshold), np.log1p(5*voltage_threshold)], [0, num_y_pos]])[0]
         hist_arrays += hist_plot
 
     log_hist_arrays = np.log1p(hist_arrays) #Take log counts 
@@ -272,36 +274,44 @@ def make_histograms(batch_id, reader, output_directory, output_directory_spikes,
 
     # print("HISTOGRAMS CREATED !!!!!!!!!")
 
-def register_data(batch_id, reader, output_directory, estimated_displacement, geomarray, out_dtype):
+def register_data(batch_id, filtered_directory, output_directory, estimated_displacement, geomarray, out_dtype):
     """Create histograms for each batches, before registration
     Parameters
     ----------
     estimated_displacement : np array containing displacements 
     """
     logger = logging.getLogger(__name__)
-    ts = reader.read_data_batch(batch_id, add_buffer=False)
+#     ts = reader.read_data_batch(batch_id, add_buffer=False)
+    fname = os.path.join(
+        filtered_directory,
+        "filtered_{}.npy".format(
+            str(batch_id).zfill(6)))
+    ts = np.load(fname)
     ts = ts.T
 
+    
+    new_data = griddata(geomarray, ts, geomarray + [0, estimated_displacement[batch_id]], fill_value = 0)
     # Griddata
-    # new_data = griddata(geomarray, ts, geomarray + [0, estimated_displacement[batch_id]], method = 'linear', fill_value = 0)
 
-    mat = np.zeros((385, 385)) #displacement = 0
-    mat_bis = np.zeros((385, 385))
-    sigma = 20*np.sqrt(2)
-    # if batch_id == 10:
-    #     print("GPR")
-    for i in range(385):
-        mat_bis[:, i] = np.exp(-np.sum((geomarray-geomarray[i])**2, axis = 1)/(2*sigma**2))
-        mat[:, i] = np.exp(-np.sum((geomarray-geomarray[i]+[0, estimated_displacement[batch_id]])**2, axis=1)/(2*sigma**2))
-    mat /= mat.sum(0)
-    mat[np.where(np.isnan(mat))]=0
-    mat_bis /= mat_bis.sum(0)
-    mat_bis[np.where(np.isnan(mat_bis))]=0
-
-    inv_mat = np.linalg.inv(mat_bis)
-    mat = np.matmul(mat, inv_mat)
-    mat[np.where(np.isnan(mat))]=0
-    new_data = np.matmul(mat, ts)
+#     mat = np.zeros((geomarray.shape[0], geomarray.shape[0])) #displacement = 0
+#     mat_bis = np.zeros((geomarray.shape[0], geomarray.shape[0]))
+#     sigma = 20*np.sqrt(2)
+#     sigma = 10
+#     # if batch_id == 10:
+#     #     print("GPR")
+#     for i in range(geomarray.shape[0]):
+#         mat_bis[:, i] = np.exp(-np.sum((geomarray-geomarray[i])**2, axis = 1)/(2*sigma**2))
+#         mat[:, i] = np.exp(-np.sum((geomarray-geomarray[i]+[0, estimated_displacement[batch_id]])**2, axis=1)/(2*sigma**2))
+#     mat += 0.01 * np.eye(mat.shape[0])
+#     mat_bis += 0.01 * np.eye(mat.shape[0])
+# #     mat[np.where(np.isnan(mat))]=0
+# #     mat_bis /= mat_bis.sum(0)
+# #     mat_bis[np.where(np.isnan(mat_bis))]=0
+# # 
+#     inv_mat = np.linalg.inv(mat_bis)
+#     mat = np.matmul(mat, inv_mat)
+# #     mat[np.where(np.isnan(mat))]=0
+#     new_data = np.matmul(mat, ts)
     # save
     fname = os.path.join(
         output_directory,
@@ -315,7 +325,7 @@ def register_data(batch_id, reader, output_directory, estimated_displacement, ge
         fname = 'Matrix_After_Registration.npy'
         np.save(fname, new_data.T.astype(out_dtype))
 
-def make_histograms_gpu(batch_id, reader, output_directory, output_directory_spikes, num_chan, sample_rate, template_time, voltage_threshold, 
+def make_histograms_gpu(batch_id, filtered_directory, output_directory, output_directory_spikes, num_chan, sample_rate, template_time, voltage_threshold, 
     length_um, num_bins, num_y_pos, quantile_comp, br_quantile, iter_quantile, neighboring_chan, M):
     """Create histograms for each batches, before registration
     Parameters
@@ -332,43 +342,69 @@ def make_histograms_gpu(batch_id, reader, output_directory, output_directory_spi
     """
     logger = logging.getLogger(__name__)
 #     print("GPU")
-    ts = torch.from_numpy(reader.read_data_batch(batch_id, add_buffer=False)).float().cuda()
+#     ts = torch.from_numpy(reader.read_data_batch(batch_id, add_buffer=False)).float().cuda()
     
     fname = os.path.join(
         output_directory,
         "histogram_{}.npy".format(
             str(batch_id).zfill(6)))
     if os.path.exists(fname):
-        return
+        return np.load(fname)[None]
+    ts = torch.from_numpy(np.load(os.path.join(
+        filtered_directory,
+        "filtered_{}.npy".format(
+            str(batch_id).zfill(6))))).float().cuda()
 
     ### Detect spikes by simple thresholding ###
     num_timesteps = ts.shape[1]
 
     ### Make histograms ###
 
-    hist_arrays = np.zeros((num_bins, num_y_pos))
     mp2d= torch.nn.MaxPool2d(kernel_size = [template_time, 1], stride = [quantile_comp,1])
     ptp_sliding = mp2d(ts[None])[0] + mp2d(-ts[None])[0]
+    
+#     print("what the fuck")
     
     for k in range(iter_quantile): 
         quantile_s = torch.kthvalue(ptp_sliding, int(br_quantile * ptp_sliding.shape[1]), dim = 1, keepdims = True)[0] #size num_timepoints
         ptp_sliding = torch.nn.functional.relu(ptp_sliding - quantile_s)
         quantile_t = torch.kthvalue(ptp_sliding, int(br_quantile * ptp_sliding.shape[0]), dim = 0, keepdims = True)[0] #size num_channels
-        ptp_sliding = torch.nn.functional.relu(ptp_sliding - quantile_t)    
+        ptp_sliding = torch.nn.functional.relu(ptp_sliding - quantile_t)
         
+    if batch_id == 5:
+        print(ptp_sliding.sum())
     ptp_sliding = ptp_sliding.cpu().numpy()
-    spike_times = np.logical_and(ptp_sliding.max(1) >= voltage_threshold, ptp_sliding.max(1) <= 10 *voltage_threshold)
-    electrode_ptp_int = np.log1p(np.matmul(ptp_sliding[spike_times],M))
+    electrode_ptp_int = np.matmul(ptp_sliding,M)
+    if batch_id == 5:
+        print(electrode_ptp_int.sum(), np.arange(num_y_pos).max(), electrode_ptp_int.shape[0])
     
-    bins = np.tile(np.arange(num_y_pos, dtype = int), electrode_ptp_int.shape[0])
-    hist_arrays = histogram2d(electrode_ptp_int.ravel(), bins, bins= (20, num_y_pos), range = [[np.log1p(voltage_threshold), np.log1p(10*voltage_threshold)], [0, num_y_pos]])
-    log_hist_arrays = np.log1p(hist_arrays) #Take log counts 
-
+    bins = np.tile(np.arange(num_y_pos), electrode_ptp_int.shape[0])
+    
+    hist_arrays = histogram2d(electrode_ptp_int.ravel(), bins, bins= (num_bins, num_y_pos), range = [[voltage_threshold, 4*voltage_threshold], [0, num_y_pos]])
+    if batch_id == 5:
+        print(hist_arrays.ptp(), voltage_threshold, 2.5*voltage_threshold, num_bins, num_y_pos)
+    
+    hist_arrays /= (hist_arrays.sum(0, keepdims = True)+1e-6)
+    if batch_id == 5:
+        print(hist_arrays.ptp())
+    log_hist_arrays = np.log1p(hist_arrays) #Take log counts
     ####### Save histogram arrays #######
     np.save(fname, log_hist_arrays)
+    
+    return(log_hist_arrays[None])
 
     # print("HISTOGRAMS CREATED !!!!!!!!!")
+def calc_displacement(displacement, niter = 50):
+    p = np.zeros(displacement.shape[0])
+    pprev = p
+    for i in range(niter):
 
+        p = (displacement.sum(1) - (p-p.sum()))/(displacement.shape[0]-1)
+        if np.allclose(p - p[0], pprev - pprev[0]):
+            break
+        else:
+            pprev = p
+    return p
 
 def merge_filtered_files(filtered_location, output_directory, delete=True, op = 'standardize'):
 
@@ -386,7 +422,9 @@ def merge_filtered_files(filtered_location, output_directory, delete=True, op = 
 
     f = open(f_out, 'wb')
     for fname in filenames_sorted:
-        res = np.load(os.path.join(filtered_location, fname))
+        if '.ipynb' in fname:
+            continue
+        res = np.load(os.path.join(filtered_location, fname)).astype('float32')
         res.tofile(f)
         if delete==True:
             os.remove(os.path.join(filtered_location, fname))
